content/canvas/test/test_2d.line.join.invalid.html

<!DOCTYPE HTML>
<title>Canvas test: 2d.line.join.invalid - bug 401788</title>
<!-- Testing: Setting lineJoin to invalid values is ignored -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

var _thrown_outer = false;
try {

ctx.lineJoin = 'bevel'
ok(ctx.lineJoin === 'bevel', "ctx.lineJoin === 'bevel'");

ctx.lineJoin = 'bevel';
ctx.lineJoin = 'invalid';
ok(ctx.lineJoin === 'bevel', "ctx.lineJoin === 'bevel'");

ctx.lineJoin = 'bevel';
ctx.lineJoin = 'ROUND';
ok(ctx.lineJoin === 'bevel', "ctx.lineJoin === 'bevel'");

ctx.lineJoin = 'bevel';
ctx.lineJoin = 'round\0';
ok(ctx.lineJoin === 'bevel', "ctx.lineJoin === 'bevel'");

ctx.lineJoin = 'bevel';
ctx.lineJoin = 'round ';
ok(ctx.lineJoin === 'bevel', "ctx.lineJoin === 'bevel'");

ctx.lineJoin = 'bevel';
ctx.lineJoin = "";
ok(ctx.lineJoin === 'bevel', "ctx.lineJoin === 'bevel'");

ctx.lineJoin = 'bevel';
ctx.lineJoin = 'butt';
ok(ctx.lineJoin === 'bevel', "ctx.lineJoin === 'bevel'");

} catch (e) {
    _thrown_outer = true;
}
todo(!_thrown_outer, 'should not throw exception');

SimpleTest.finish();

});
</script>
